+++ src/test/java/roomescape/test/fake/FakeH2ReservationRepository.java
@@ -9,14 +9,10 @@ import java.util.concurrent.atomic.AtomicLong;
 import roomescape.domain.Reservation;
 import roomescape.repository.ReservationRepository;
 
-public class FakeReservationRepository extends ReservationRepository {
+public class FakeH2ReservationRepository implements ReservationRepository {
 
     private final Map<Long, Reservation> reservations = new ConcurrentHashMap<>();
     private final AtomicLong index = new AtomicLong(1);
-
-    public FakeReservationRepository() {
-        super(null);
-    }
 
     @Override
     public List<Reservation> findAll() {
